feat: 게시물 리스트 검색로직 수정 (#328)
