+++ frontend/src/components/Apply/index.jsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 
 import { bookmarkApi, feedlistApi } from "../../utils/feedApi";
 import { useRecoilState, useRecoilValue } from "recoil";
@@ -16,78 +16,54 @@ import SearchBar from "./SearchBar";
 import ListContainer from "./ListContainer";
 
 import * as S from "./style";
+import { useHistory } from "react-router";
 
-export default function ApplyComponent() {
+export default function ApplyComponent({ page }) {
+  const history = useHistory();
   const user = useRecoilValue(userState);
   const [filter, setFilter] = useRecoilState(filterState);
   const [school, setSchool] = useRecoilState(schoolState);
   const [category, setCategory] = useRecoilState(categoryState);
   const [date, setDate] = useRecoilState(dateState);
   const [feedList, setFeedList] = useState();
-  const [page, setPage] = useState(1);
   const [totalElements, setTotalElements] = useState(0);
 
-  const getFeedList = useCallback(
-    async (page, schoolId, date, filter, category) => {
-      await feedlistApi(page, schoolId, date, filter, category, user.isLogin)
+  const search = useCallback(
+    (schoolId, page) => {
+      feedlistApi(page, schoolId, date, filter, category, user.isLogin)
         .then((res) => {
           setFeedList(res.data.data.content);
           setTotalElements(res.data.data.totalElements);
+          history.push(`/apply?${page}`);
         })
         .catch((err) => console.log(err));
     },
-    [user.isLogin]
+    [user.isLogin, category, date, filter, history]
   );
 
-  const postBookmark = async (postId) => {
-    await bookmarkApi(postId)
+  const postBookmark = (postId) => {
+    bookmarkApi(postId)
       .then()
       .catch((err) => console.log(err));
   };
 
-  const search = useCallback(
-    (e) => {
-      if (typeof e === "number") {
-        setPage(e);
-        if (school.schoolId) {
-          getFeedList(e, school.schoolId, date, filter, category);
-        } else if (user.schoolId) {
-          getFeedList(e, user.schoolId, date, filter, category);
-        } else {
-          alert("학교를 선택해주세요");
-        }
-      } else {
-        if (school.schoolId) {
-          getFeedList(page, school.schoolId, date, filter, category);
-        } else if (user.schoolId) {
-          getFeedList(page, user.schoolId, date, filter, category);
-        } else {
-          alert("학교를 선택해주세요");
-        }
-      }
-    },
-    [category, date, filter, getFeedList, page, school.schoolId, user.schoolId]
-  );
+  useEffect(() => {
+    if (page) search(school.schoolId, page);
+  }, [search, page, school.schoolId]);
 
   return (
     <S.ListPageContainer>
       <SearchBar
-        feedList={feedList}
-        search={search}
-        filter={filter}
-        setFilter={setFilter}
         school={school}
         setSchool={setSchool}
-        category={category}
-        setCategory={setCategory}
         date={date}
         setDate={setDate}
+        search={search}
       />
       {feedList && (
         <>
           <Filter
             feedList={feedList}
-            search={search}
             filter={filter}
             setFilter={setFilter}
             category={category}
@@ -104,8 +80,7 @@ export default function ApplyComponent() {
             <Pagination
               page={page}
               count={totalElements}
-              setPage={setPage}
-              onChange={search}
+              onChange={(e) => history.push(`/apply?${e}`)}
             />
           )}
         </>
